requirements.txt
# Serverless RAG Project - Python Dependencies

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# OpenAI
openai>=1.12.0

# Pinecone
pinecone-client>=3.0.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# pgvector support (optional)
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
moto>=5.0.0  # AWS mocking

# Development
python-dotenv>=1.0.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

# Utilities
requests>=2.31.0
tenacity>=8.2.0  # Retry logic
